+++ src/main/java/net/geforcemods/securitycraft/SCClientEventHandler.java
@@ -1,18 +1,23 @@
 package net.geforcemods.securitycraft;
 
+import com.mojang.blaze3d.pipeline.RenderTarget;
 import com.mojang.blaze3d.platform.Window;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.blaze3d.vertex.VertexConsumer;
 
 import net.geforcemods.securitycraft.api.IExplosive;
 import net.geforcemods.securitycraft.api.ILockable;
 import net.geforcemods.securitycraft.api.IOwnable;
+import net.geforcemods.securitycraft.blockentities.BlockChangeDetectorBlockEntity;
+import net.geforcemods.securitycraft.blockentities.BlockChangeDetectorBlockEntity.ChangeEntry;
 import net.geforcemods.securitycraft.blockentities.SecurityCameraBlockEntity;
 import net.geforcemods.securitycraft.blocks.DisguisableBlock;
 import net.geforcemods.securitycraft.blocks.SecurityCameraBlock;
 import net.geforcemods.securitycraft.entity.Sentry;
 import net.geforcemods.securitycraft.entity.camera.SecurityCamera;
 import net.geforcemods.securitycraft.items.SonicSecuritySystemItem;
+import net.geforcemods.securitycraft.misc.BlockEntityTracker;
 import net.geforcemods.securitycraft.misc.KeyBindings;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.misc.SCSounds;
@@ -24,6 +29,8 @@ import net.minecraft.client.Options;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiComponent;
 import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
@@ -39,6 +46,7 @@ import net.minecraft.world.level.ClipContext;
 import net.minecraft.world.level.ClipContext.Block;
 import net.minecraft.world.level.ClipContext.Fluid;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.HitResult;
@@ -46,6 +54,7 @@ import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.InputEvent;
 import net.minecraftforge.client.event.RenderHandEvent;
+import net.minecraftforge.client.event.RenderWorldLastEvent;
 import net.minecraftforge.client.event.ScreenshotEvent;
 import net.minecraftforge.client.gui.ForgeIngameGui;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -58,6 +67,31 @@ public class SCClientEventHandler {
 	public static final ResourceLocation NIGHT_VISION = new ResourceLocation("textures/mob_effect/night_vision.png");
 	private static final ItemStack REDSTONE = new ItemStack(Items.REDSTONE);
 	private static final Component REDSTONE_NOTE = Utils.localize("gui.securitycraft:camera.toggleRedstoneNote");
+
+	@SubscribeEvent
+	public static void onRenderLevelStage(RenderWorldLastEvent event) {
+		Minecraft mc = Minecraft.getInstance();
+		Vec3 camPos = mc.gameRenderer.getMainCamera().getPosition();
+		PoseStack pose = event.getMatrixStack();
+		Level level = mc.level;
+
+		for (BlockPos bcdPos : BlockEntityTracker.BLOCK_CHANGE_DETECTOR.getTrackedBlockEntities(level)) {
+			BlockEntity be = level.getBlockEntity(bcdPos);
+
+			if (be instanceof BlockChangeDetectorBlockEntity bcd && bcd.isShowingHighlights() && bcd.getOwner().isOwner(mc.player)) {
+				for (ChangeEntry changeEntry : bcd.getFilteredEntries()) {
+					BlockPos pos = changeEntry.pos();
+
+					pose.pushPose();
+					pose.translate(pos.getX() - camPos.x, pos.getY() - camPos.y, pos.getZ() - camPos.z);
+					ClientUtils.renderBoxInLevel(BCDBuffer.INSTANCE, pose.last().pose(), 0, 1, 0, 1, 1, bcd.getColor());
+					pose.popPose();
+				}
+			}
+		}
+
+		mc.renderBuffers().bufferSource().endBatch();
+	}
 
 	@SubscribeEvent
 	public static void onScreenshot(ScreenshotEvent event) {
@@ -145,9 +179,8 @@ public class SCClientEventHandler {
 			else
 				gui.blit(pose, 12, 3, 90, 0, 12, 11);
 		}
-		else {
+		else
 			Minecraft.getInstance().getItemRenderer().renderAndDecorateItem(REDSTONE, 10, 0);
-		}
 	}
 
 	public static void hotbarBindOverlay(ForgeIngameGui gui, PoseStack pose, float partialTicks, int width, int height) {
@@ -245,6 +278,42 @@ public class SCClientEventHandler {
 			if (uCoord != 0) {
 				RenderSystem._setShaderTexture(0, BEACON_GUI);
 				GuiComponent.blit(pose, Minecraft.getInstance().getWindow().getGuiScaledWidth() / 2 - 90 + (hand == InteractionHand.MAIN_HAND ? player.getInventory().selected * 20 : (mc.options.mainHand == HumanoidArm.LEFT ? 189 : -29)), Minecraft.getInstance().getWindow().getGuiScaledHeight() - 22, uCoord, 219, 21, 22, 256, 256);
+			}
+		}
+	}
+
+	private static enum BCDBuffer implements MultiBufferSource {
+		INSTANCE;
+
+		private final RenderType overlayLines = new OverlayLines(RenderType.lines());
+
+		@Override
+		public VertexConsumer getBuffer(RenderType renderType) {
+			return Minecraft.getInstance().renderBuffers().bufferSource().getBuffer(overlayLines);
+		}
+
+		private static class OverlayLines extends RenderType {
+			private final RenderType normalLines;
+
+			private OverlayLines(RenderType normalLines) {
+				super("overlay_lines", normalLines.format(), normalLines.mode(), normalLines.bufferSize(), normalLines.affectsCrumbling(), normalLines.sortOnUpload, normalLines::setupRenderState, normalLines::clearRenderState);
+				this.normalLines = normalLines;
+			}
+
+			@Override
+			public void setupRenderState() {
+				normalLines.setupRenderState();
+
+				RenderTarget renderTarget = Minecraft.getInstance().levelRenderer.entityTarget();
+
+				if (renderTarget != null)
+					renderTarget.bindWrite(false);
+			}
+
+			@Override
+			public void clearRenderState() {
+				Minecraft.getInstance().getMainRenderTarget().bindWrite(false);
+				normalLines.clearRenderState();
 			}
 		}
 	}