+++ src/main/java/net/geforcemods/securitycraft/util/ClientUtils.java
@@ -8,6 +8,8 @@ import com.mojang.blaze3d.vertex.BufferUploader;
 import com.mojang.blaze3d.vertex.DefaultVertexFormat;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.Tesselator;
+import com.mojang.blaze3d.vertex.VertexConsumer;
+import com.mojang.blaze3d.vertex.VertexFormat;
 import com.mojang.blaze3d.vertex.VertexFormat.Mode;
 import com.mojang.math.Matrix4f;
 
@@ -15,6 +17,8 @@ import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 
@@ -87,5 +91,164 @@ public class ClientUtils {
 		bufferBuilder.vertex(m4f, moduleLeft, moduleTop, 0).color(1.0F, 1.0F, 1.0F, alpha).uv(0, 0).endVertex();
 		bufferBuilder.end();
 		BufferUploader.end(bufferBuilder);
+	}
+
+	public static void renderBoxInLevel(MultiBufferSource buffer, Matrix4f positionMatrix, int minX, int maxX, int minZ, int maxZ, int height, int rgbColor) {
+		VertexConsumer builder = buffer.getBuffer(RenderType.lines());
+		int r = rgbColor >> 16 & 255;
+		int g = rgbColor >> 8 & 255;
+		int b = rgbColor & 255;
+
+		//bottom lines
+		builder.vertex(positionMatrix, minX, 0.0F, minZ).color(r, g, b, 255).normal(0.0F, 0.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, 0.0F, minZ).color(r, g, b, 255).normal(0.0F, 0.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, minX, 0.0F, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, 0.0F, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, minX, 0.0F, minZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, minX, 0.0F, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, 0.0F, minZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, 0.0F, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		//top lines
+		builder.vertex(positionMatrix, minX, height, minZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, height, minZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, minX, height, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, height, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, minX, height, minZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, minX, height, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, height, minZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, height, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		//corner edge lines
+		builder.vertex(positionMatrix, minX, 0.0F, minZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, minX, height, minZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, minX, 0.0F, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, minX, height, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, 0.0F, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, height, maxZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, 0.0F, minZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+		builder.vertex(positionMatrix, maxX, height, minZ).color(r, g, b, 255).normal(1.0F, 1.0F, 1.0F).endVertex();
+	}
+
+	public static void fillHorizontalGradient(PoseStack pose, int zLevel, int left, int top, int right, int bottom, int fromColor, int toColor) {
+		float fromAlpha = (fromColor >> 24 & 255) / 255.0F;
+		float fromRed = (fromColor >> 16 & 255) / 255.0F;
+		float fromGreen = (fromColor >> 8 & 255) / 255.0F;
+		float fromBlue = (fromColor & 255) / 255.0F;
+		float toAlpha = (toColor >> 24 & 255) / 255.0F;
+		float toRed = (toColor >> 16 & 255) / 255.0F;
+		float toGreen = (toColor >> 8 & 255) / 255.0F;
+		float toBlue = (toColor & 255) / 255.0F;
+		Matrix4f mat = pose.last().pose();
+		Tesselator tessellator = Tesselator.getInstance();
+		BufferBuilder buffer = tessellator.getBuilder();
+
+		RenderSystem.enableDepthTest();
+		RenderSystem.disableTexture();
+		RenderSystem.enableBlend();
+		RenderSystem.defaultBlendFunc();
+		RenderSystem.setShader(GameRenderer::getPositionColorShader);
+		buffer.begin(VertexFormat.Mode.QUADS, DefaultVertexFormat.POSITION_COLOR);
+		buffer.vertex(mat, right, top, zLevel).color(toRed, toGreen, toBlue, toAlpha).endVertex();
+		buffer.vertex(mat, left, top, zLevel).color(fromRed, fromGreen, fromBlue, fromAlpha).endVertex();
+		buffer.vertex(mat, left, bottom, zLevel).color(fromRed, fromGreen, fromBlue, fromAlpha).endVertex();
+		buffer.vertex(mat, right, bottom, zLevel).color(toRed, toGreen, toBlue, toAlpha).endVertex();
+		tessellator.end();
+		RenderSystem.disableBlend();
+		RenderSystem.enableTexture();
+	}
+
+	public static int HSBtoRGB(float hue, float saturation, float brightness) {
+		int r = 0;
+		int g = 0;
+		int b = 0;
+
+		if (saturation == 0)
+			r = g = b = (int) (brightness * 255.0F + 0.5F);
+		else {
+			float h = (hue - (float) Math.floor(hue)) * 6.0F;
+			float f = h - (float) Math.floor(h);
+			float p = brightness * (1.0F - saturation);
+			float q = brightness * (1.0F - saturation * f);
+			float t = brightness * (1.0F - (saturation * (1.0F - f)));
+
+			switch ((int) h) {
+				case 0:
+					r = (int) (brightness * 255.0F + 0.5F);
+					g = (int) (t * 255.0F + 0.5F);
+					b = (int) (p * 255.0F + 0.5F);
+					break;
+				case 1:
+					r = (int) (q * 255.0F + 0.5F);
+					g = (int) (brightness * 255.0F + 0.5F);
+					b = (int) (p * 255.0F + 0.5F);
+					break;
+				case 2:
+					r = (int) (p * 255.0F + 0.5F);
+					g = (int) (brightness * 255.0F + 0.5F);
+					b = (int) (t * 255.0F + 0.5F);
+					break;
+				case 3:
+					r = (int) (p * 255.0F + 0.5F);
+					g = (int) (q * 255.0F + 0.5F);
+					b = (int) (brightness * 255.0F + 0.5F);
+					break;
+				case 4:
+					r = (int) (t * 255.0F + 0.5F);
+					g = (int) (p * 255.0F + 0.5F);
+					b = (int) (brightness * 255.0F + 0.5F);
+					break;
+				case 5:
+					r = (int) (brightness * 255.0F + 0.5F);
+					g = (int) (p * 255.0F + 0.5F);
+					b = (int) (q * 255.0F + 0.5F);
+					break;
+			}
+		}
+
+		return 0xFF000000 | (r << 16) | (g << 8) | (b << 0);
+	}
+
+	public static float[] RGBtoHSB(int r, int g, int b) {
+		float hue, saturation, brightness;
+		float[] hsb = new float[3];
+		int cmax = (r > g) ? r : g;
+		int cmin = (r < g) ? r : g;
+
+		if (b > cmax)
+			cmax = b;
+
+		if (b < cmin)
+			cmin = b;
+
+		brightness = cmax / 255.0F;
+
+		if (cmax != 0)
+			saturation = ((float) (cmax - cmin)) / ((float) cmax);
+		else
+			saturation = 0;
+
+		if (saturation == 0)
+			hue = 0;
+		else {
+			float redc = (float) (cmax - r) / (float) (cmax - cmin);
+			float greenc = (float) (cmax - g) / (float) (cmax - cmin);
+			float bluec = (float) (cmax - b) / (float) (cmax - cmin);
+
+			if (r == cmax)
+				hue = bluec - greenc;
+			else if (g == cmax)
+				hue = 2.0F + redc - bluec;
+			else
+				hue = 4.0F + greenc - redc;
+
+			hue = hue / 6.0F;
+
+			if (hue < 0)
+				hue = hue + 1.0F;
+		}
+
+		hsb[0] = hue;
+		hsb[1] = saturation;
+		hsb[2] = brightness;
+		return hsb;
 	}
 }